EZP-29140: Registration
