+++ lib/Data/Mapper/UserRegisterMapper.php
@@ -7,78 +7,12 @@
  */
 namespace EzSystems\RepositoryForms\Data\Mapper;
 
-use eZ\Publish\API\Repository\Values\Content\Field;
-use EzSystems\RepositoryForms\Data\Content\FieldData;
-use EzSystems\RepositoryForms\Data\User\UserRegisterData;
-use EzSystems\RepositoryForms\User\RegistrationContentTypeLoader;
-use EzSystems\RepositoryForms\User\RegistrationGroupLoader;
-use Symfony\Component\OptionsResolver\OptionsResolver;
+use EzSystems\EzPlatformUser\Form\DataMapper\UserRegisterMapper as BaseUserRegisterMapper;
 
 /**
  * Form data mapper for user creation.
+ * @deprecated Deprecated in 1.1 and will be removed in 2.0. Please use EzSystems\EzPlatformUser\Form\DataMapper\UserRegisterMapper instead.
  */
-class UserRegisterMapper
+class UserRegisterMapper extends BaseUserRegisterMapper
 {
-    /**
-     * @var RegistrationContentTypeLoader
-     */
-    private $contentTypeLoader;
-
-    /**
-     * @var RegistrationGroupLoader
-     */
-    private $parentGroupLoader;
-
-    /**
-     * @var array
-     */
-    private $params;
-
-    public function __construct(RegistrationContentTypeLoader $contentTypeLoader, RegistrationGroupLoader $registrationGroupLoader)
-    {
-        $this->contentTypeLoader = $contentTypeLoader;
-        $this->parentGroupLoader = $registrationGroupLoader;
-    }
-
-    public function setParam($name, $value)
-    {
-        $this->params[$name] = $value;
-    }
-
-    /**
-     * @return UserRegisterData
-     */
-    public function mapToFormData()
-    {
-        $resolver = new OptionsResolver();
-        $this->configureOptions($resolver);
-        $this->params = $resolver->resolve($this->params);
-
-        $contentType = $this->contentTypeLoader->loadContentType();
-
-        $data = new UserRegisterData([
-            'contentType' => $contentType,
-            'mainLanguageCode' => $this->params['language'],
-            'enabled' => true,
-        ]);
-        $data->addParentGroup($this->parentGroupLoader->loadGroup());
-
-        foreach ($contentType->fieldDefinitions as $fieldDef) {
-            $data->addFieldData(new FieldData([
-                'fieldDefinition' => $fieldDef,
-                'field' => new Field([
-                    'fieldDefIdentifier' => $fieldDef->identifier,
-                    'languageCode' => $this->params['language'],
-                ]),
-                'value' => $fieldDef->defaultValue,
-            ]));
-        }
-
-        return $data;
-    }
-
-    private function configureOptions(OptionsResolver $optionsResolver)
-    {
-        $optionsResolver->setRequired('language');
-    }
 }
